Fixing script commands that dont work on windows.
Former-commit-id: 12fb3e16ae53b18bb83d32756ddf55c0726eacca

+++ install/Shared/script/cx/cxTestRunner.py
@@ -121,7 +121,8 @@ TimeOut: %d
         PrintFormatter.printInfo('Run catch wit tag [%s], results to %s' % (tag, outfile))
         shell.changeDir(path)
         #shell.run('rm -rf %s' % outfile)
-        os.remove(outfile)
+        if os.path.exists(filename):
+            os.remove(outfile)
         shell.run('%s/Catch [%s] --reporter junit --out %s' % (path, tag, outfile), ignoreFailure=True)
         #shell.run('%s/Catch [%s] --out %s' % (path, tag, outfile))
